Modified desired table logic
* It was the case the the desired table was computed by
taking the last digit of the group, and saying that the
table on that position in the HTML was desired. This
does not work for cases where the last digit does
not represent the actual index of the desired table.

* The current implementation takes advantage of the
label that always precedes the tables. If you want table
248/2, then it will look for that label, and take the next
table it finds from there.